requirements.txt
line-bot-sdk==1.8.0
Django==2.2
openai


